refactor Vector4 implementation code from PixelOperation<T> to Vector4Converters
