+++ tests/ImageSharp.Tests/PixelFormats/PixelConverterTests.cs
@@ -1,4 +1,5 @@
 ﻿using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp.PixelFormats.Utils;
 
 using Xunit;
